Move report_builder.py to be a classmethod in Report. Use 'lower' instead of linear interpolation for numpy percentile
